Update README.md
This lab will port the application from the last two labs, a 1 Msps digital oscilloscope, to a RTOS (TI-RTOS).

To complete all labs in this course, you will need to submit your source for that lab, and demonstrate the lab in operation to one of the course staff (a TA or tutor). You will be graded on the completion of the steps in this lab document. After you have received a signoff, follow the instructions at the end of the document to submit your source code.

Objectives

Port an existing real-time application to an RTOS

Use fundamental RTOS objects: Task, Hwi, Clock, Semaphore, Mailbox

Deal with shared data and other inter-task communication